chore: Update home_view to return JSON response
This commit updates the home_view function in views.py to return a JSON response instead of rendering the 'home.html' template. It adds the necessary import statement for the JsonResponse class from django.http and modifies the return statement to return a JSON object with a 'message' key and value of 'Home!'. This change improves the functionality of the home page by returning data in a more structured format.